Add some heuristic to process the alpha channel on Windows
When we get the clipboard bitmap from Windows, we're completely lost
about the content of the alpha channel (it can be anything). So if all
pixels contains alpha = 0, it means that the alpha isn't correct. If
they are valid premultiplied-alpha values, the values will be
interpreted as premultiplied-alpha. In other case alpha will be
interpreted as straight alpha (I'm not sure if this is correct, I think
Chrome in this case just ignore alpha and make it =255 for every pixel).

In any case, with this change:
* Screenshots are working again (we can get the clipboard image from a
  Print Screen correctly again, because it looks like alpha=0 for all
  pixels in this case)
* We can copy an image with alpha channel (e.g. a png image) from
  Chrome (right-click > Copy Image) and we'll be able to get the alpha
  information correctly.