Initial experimental X11 support
This first version is an experimental version to provide text clipboard
content to other processes on X11. i.e. We can "paste" text content in
the clipboard. Anyway this initial implementation isn't getting
clipboard content from other processes yet.

Note: We're using xcb library because Xlib is not multi-threading
friendly and we need to run a event loop in the background.