release m_mutex in get_data_from_selection_owner()
Release m_mutex when get_data_from_selection_owner() returns.
Otherwise the mutex gets unlocked twice, first in the destructor
of std::unique_lock and again in lock::impl::~impl().

Found when trying to run this code on OpenBSD, which calls
abort() in pthread_mutex_unlock() with an unlocked mutex.