[osx] Fix memory leaks using -fobjc-arc correctly
The -fobjc-arc flags must be specified for the compiler, not in the
linker. This generates a lot of memory leaks when we copy content into
the clipboard (just testing copy/put_image examples with a for loop to
repeat the action can easily show this leak).

This might be the reason of some memory leaks reported on the macOS
port of Aseprite: https://github.com/aseprite/aseprite/issues/4251